rag-service/pyproject.toml
[project]
name = "rag-service"
version = "0.1.0"
description = "RAG Service for document retrieval and embeddings"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.0",
    "asyncpg>=0.29.0",
    "llama-index-core>=0.10.0",
    "llama-index-vector-stores-postgres>=0.1.0",
    "sentence-transformers>=2.2.0"
]
